configure.ac: fix incorrect bashism
The POSIX test command only accepts single-equals. The use of double
equals is a bashism that is exactly equal to using single-equals, and
must never be used. It provides no advantages in any scenario, and it
messes with muscle memory when one has to write portable POSIX sh
scripts.

In this case, configure scripts are run via /bin/sh, which may or may
not be bash and has to assume POSIX compatibility. So this produces the
following Gentoo build warning:

 * QA Notice: Abnormal configure code
 *
 * ./configure: 22449: test: xno: unexpected operator

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>